requirements.txt
altgraph==0.17.4
numpy==2.1.3
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pygame==2.6.1
pygame-ce==2.5.2
pygame_gui==0.6.12
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-i18n==0.3.9
pywin32-ctypes==0.2.3
